requirements.txt
dnspython==2.5.0
exceptiongroup==1.2.0
Faker==23.1.0
iniconfig==2.0.0
mysql-connector-python==8.2.0
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pluggy==1.4.0
protobuf==4.21.12
psycopg2==2.9.9
pymongo==4.6.1
pytest==8.0.0
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
tomli==2.0.1
tzdata==2024.1
